Updated vae_compression and logbook
Changed vae_compression to include proper imports and took a different approach to implementing the model as the previous one was not working. It does not fully work yet but I hope to have the new error fixed when possible. Updated logbook to reflect this as well as include some needed links.